fix: swap issue
