+++ src/utils/bridge.ts
@@ -1,6 +1,4 @@
-import { post, isSwap, swapContractAddressCheck, getGasPrice, isNativeToken } from ".";
-import { addChainData, ChainBackendNames, CHAIN_ID_HEX_TO_ENUM_MAPPING, contractABI, CONTRACT_DECIMAL_TO_ETHER_UNITS } from "../constants/server";
-import { switchNetwork } from "../core";
+import { post, isSwap, swapContractAddressCheck, isNativeToken, checkAllowance, getSwapAllowanceApproval } from ".";
 
 declare let globalThis: any;
 
@@ -58,8 +56,10 @@ export const onGetQuote = async () => {
                   contractData: allowanceResp.contractData,
                   gasPrice: allowanceResp.gasPrice,
                 };
-                bridgeSubmitButton!.disabled = false;
-                bridgeSubmitButton!.classList.remove("disabled-button");
+                if(await getSwapAllowanceApproval()) {
+                  bridgeSubmitButton!.disabled = false;
+                  bridgeSubmitButton!.classList.remove("disabled-button");
+                }
               } else {
                 globalThis.allowanceData = {
                   isAllowance: false
@@ -76,7 +76,7 @@ export const onGetQuote = async () => {
             }
           } else {
             globalThis.allowanceData = {
-              isAllowance: true
+              isAllowance: false
             };
             bridgeSubmitButton!.disabled = false;
             bridgeSubmitButton!.classList.remove("disabled-button");
@@ -131,58 +131,6 @@ export const onGetQuote = async () => {
         }
       });
   }
-}
-
-export const checkAllowance = async ({
-  chain,
-  contractAddress,
-  routerAddress,
-  amount,
-  contractDecimal,
-  isNative
-}: any) => {
-  const {
-    exchangingTokenDetail: {
-      chainDetails: {
-        chain_id
-      },
-      contractDecimals
-    },
-    cypherWalletDetails: {
-      address
-    },
-    bridgeInputDetails: {
-      tokenValueEntered
-    },
-    Cypher: {
-      Web3
-    }
-  } = globalThis;
-  await switchNetwork(chain_id);
-  const rpcEndpoint = addChainData[CHAIN_ID_HEX_TO_ENUM_MAPPING.get(chain_id)!].rpcUrls[0];
-  const web3 = new Web3(rpcEndpoint);
-  let userAddress = address;
-  if (chain === ChainBackendNames.EVMOS) {
-    userAddress = web3.utils.toChecksumAddress(userAddress);
-  }
-  const gasPrice = await getGasPrice(chain);
-  const contract = new web3.eth.Contract(contractABI, contractAddress);
-  const response = await contract.methods.allowance(userAddress, routerAddress).call();
-  const etherUnit = CONTRACT_DECIMAL_TO_ETHER_UNITS[contractDecimals];
-  const tokenAmount = web3.utils.toWei(Number(amount).toFixed(contractDecimals), etherUnit).toString();
-  if (Number(tokenAmount) > Number(response)) {
-    if (Number(amount) < 1000) amount = '1000';
-    const tokens = web3.utils.toWei((Number(amount) * 10).toFixed(contractDecimals));
-    const resp = contract.methods.approve(routerAddress, tokens).encodeABI();
-    const gasLimit = await web3.eth.estimateGas({
-      from: userAddress,
-      to: contractAddress,
-      value: isNative ? web3.utils.toWei(Number(tokenValueEntered).toFixed(contractDecimals), 'ether') : '0x0',
-      data: resp,
-    });
-    return { isError: false, isAllowance: true, contractData: resp, gasLimit: gasLimit, gasPrice };
-  }
-  return { isError: false, isAllowance: false };
 }
 
 export const onDepositFund = async (hash: string) => {
